==> analysis/05_merge_child_measle_report_updated2.Rmd <==
---
title: "Vaccination and delays in vaccination due to COVID-19"
author: "Duc Du, Thinh Ong"
date: ' 2020-01-29 (update: `r Sys.Date()`)'
output:
  workflowr::wflow_html:
    toc: true
    theme: cosmo
  pdf_document:
    toc: yes
  html_document:
    theme: cerulean
    toc: yes
    toc_float: yes
    keep_md: yes
  word_document:
    toc: yes
editor_options:
  chunk_output_type: inline
---

```{r}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE, cache.lazy = FALSE,
                      tidy.opts=list(width.cutoff=60),
                      tidy=TRUE,
                      root.dir = rprojroot::find_rstudio_root_file())

library(workflowr)
library(data.table)
library(tidyverse)
library(dplyr)
library(lubridate)
library(ggplot2)
library(gt)
library(gtsummary)
library(ggridges)

select <- dplyr::select
allp <- file.path("~", "updated_dataset")
```

# Introduction

COVID-19 creates negative impacts on other diseases and health conditions, one of them is vaccine preventable diseases (VPD). Lock-down periods may translate into delays in vaccination and delays in vaccination may result in increased risk of outbreaks of VPD. We quantified these delays in vaccination and related factors for measles and other VPD. We propose to translate these delays into outbreak risks for measles and other VPD with an epidemiological model.

# Methods

We used data from the Vietnam Vaccine Registry collected at the commune, district and provincial levels among vaccination children who undertook their vaccine shots at either public clinics (across 28 provinces in the Northern areas) under the Expanded Program on Immunization or private clinics (across 63 provinces in the whole country) over the period from 01-01-2017 to 31-12-2020. 

# Results & Discussions

```{r}
load(file.path(allp, "hepatitisb_all.Rdata"))

# Get a vector of public provinces
dtadir <- file.path("~", "sharefolder", "data", "tuann349")
publicProvinces <- list.dirs(dtadir, full.names = F, recursive = F)
publicProvinces <- stringi::stri_trans_general(publicProvinces, "any-ascii")

# Child = 28 provinces born between 2017 and 2020, vaccinated with HepBN or HepB, only the 1st shot
child_all2 <- hepatitisb_all[hepatitisb_all$province %in% publicProvinces &
                               hepatitisb_all$year >= 2017 & 
                               hepatitisb_all$year <= 2020 &
                               hepatitisb_all$vacname2 %in% c("HepBN", "HepB") &
                               hepatitisb_all$shot == 1,]
child_all2$age <- 2021 - child_all2$year
child_all2$sex <- ifelse(child_all2$sex==2, 1, child_all2$sex)
child_all2$sex <- factor(child_all2$sex, levels=c(0,1), labels=c("F","M"))
child_all_overall <- child_all2[, .N, by = .(year, month)]
child_all_province <- child_all2[, .N, by = .(year, month, province)]
```

## Characteristics of vaccination children

```{r}
t2 <- child_all2 %>% select(age, sex, province, type, type2, year) %>%
  tbl_summary(
  by=type, # split table by group
  percent = "column",
  #missing = "no" # don't list missing data separately
  ) %>%
  add_n() %>% # add column with total number of non-missing observations
  #add_p(pvalue_fun=~style_pvalue(.x,digits=2)) %>% # test for a difference between groups
  modify_header(label = "**Variable**") %>% # update the column header
  bold_labels() 
# Can't filter a tbl_summary object so I make a new one: add criterion province in public provinces
tab_child_all_sum3 <- tbl_summary(child_all2[, .(province, sex, age, type, type2, factor(year))])
# Data from GSO
birthGSO <- readRDS(file.path("~", "Vaccination_COVID", "data", "birthGSO.rds"))
birthGSO <- birthGSO[birthGSO$province %in% publicProvinces,c("province", "birth17_20")]
```

Table 1 describes the characteristics of vaccination children born from 1/1/2017 to 31/12/2020. Of total `r format(tab_child_all_sum3$N, big.mark=",")` children (`r gsub("(?<=\\()[^()]*(?=\\))(*SKIP)(*F)|.", "", as.character(tab_child_all_sum3$table_body[tab_child_all_sum3$table_body$variable=="sex" & tab_child_all_sum3$table_body$label=="F","stat_0"]), perl = T)` female) vaccinated from 2017 to 2020, `r as.character(tab_child_all_sum3$table_body[tab_child_all_sum3$table_body$variable=="type" & tab_child_all_sum3$table_body$label=="both","stat_0"])` were vaccinated at both public and private clinics, whereas `r as.character(tab_child_all_sum3$table_body[tab_child_all_sum3$table_body$variable=="type" & tab_child_all_sum3$table_body$label=="public","stat_0"])` got immunized at public clinics only and `r as.character(tab_child_all_sum3$table_body[tab_child_all_sum3$table_body$variable=="type" & tab_child_all_sum3$table_body$label=="private","stat_0"])` got immunized at private clinics only. There was an increase in the total number of children vaccinated at both public clinics and private clinics over 4 years. Similar trend was observed for those vaccinated at private clinics only from 2017 to 2019 then a bit drop in 2020 while the total number of children vaccinated at public clinics only decreased over years (Table 1).

```{r}
tab_child_all_sum3
```

Compared to GSO data:

```{r}
tb <- tab_child_all_sum3$table_body[tab_child_all_sum3$table_body$variable == "province",]
tb <- data.frame(tb)
tb <- tb[-1,c("label", "stat_0")]
tb <- merge(tb, birthGSO, by.x = "label", by.y = "province")
tb$pct <- 100 * tb$birth17_20 / sum(tb$birth17_20)
tb$gso <- paste0(format(round(tb$birth17_20), big.mark = ","), " (", sprintf("%.1f", tb$pct), "%)")
tb <- tb[,c("label", "stat_0", "gso")]
colnames(tb) <- c("province", "hepatitisb", "gso_estimation")

knitr::kable(tb)

```

```{r}
t2
```

## Child overall

## Children born by month and year

Overall trend of vaccination children born from 2017 to 2020 was shown in Figure 1. There was a steady increase in the total number of children born from February (around Tet holiday) to October, reach a peak in between September and October, then drop in the next month, annually. Similar trend was observed for each year (Figure 1). 

```{r}
ggplot(data = child_all_overall[year >= 2017 & year <= 2020], aes(x = factor(month), y = N/1000, color = factor(year))) +
  geom_rect(aes(xmin = 3, xmax = 5, ymin = -Inf, ymax = Inf), fill = "grey90", color = "grey90", alpha = 0.015) +
  geom_point(size = 3) +
  geom_line(aes(group = year), size = 1.5) +
  scale_x_discrete(name = "Month") +
  scale_y_continuous(name = "Number of children born (x 10^3)", breaks = seq(from = 0, to = 150, by = 50)) +
  scale_color_discrete(name = "Year") +
  theme_bw() +
  ggtitle("Total number of vaccination children born by month and year")
#ggsave(filename = file.path("..", "figures", "child_all_month_year.png"), width = 15, height = 10)
```

There was an increase in the total number of children born from 2017 to 2019. The number of children born then decreased in 2020 (Figure 2).

```{r}
ggplot(child_all_overall[year >= 2017 & year <= 2020], aes(month, N/1000, color = factor(year))) +
  geom_point(size=3) +
  geom_line(aes(group=year)) +
  geom_smooth () +
  facet_wrap(~ year) + 
  stat_smooth(method="loess", formula = y ~ x, colour="blue", size=1) +
  scale_x_discrete(name = "Month") +
  scale_y_continuous(name = "Number of children born (x 10^3)", breaks = seq(from = 0, to = 300, by = 50)) +
  scale_color_discrete(name = "Year") +
  theme_bw() +
  ggtitle("Total number of vaccination children born by month and year")
#ggsave(filename = file.path("..", "figures", "child_all_month_year2.png"), width = 15, height = 10)
```
## Children born by province

Total number of vaccination children born in each province by month and year from 2017 to 2020 was shown in Figure 3. In general, similar trend was observed in most of provinces (Figure 3).

```{r}
ggplot(data = child_all_province[year >= 2017 & year <= 2020], aes(x = factor(month), y = N/1000, color = factor(year))) +
  geom_rect(aes(xmin = 3, xmax = 5, ymin = -Inf, ymax = Inf), fill = "grey90", color = "grey90", alpha = 0.015) +
  geom_point() +
  geom_line(aes(group = year)) +
  scale_x_discrete(name = "Month") +
  scale_y_continuous(name = "Number of children born (x 10^3)") +
  scale_color_discrete(name = "Year") +
  facet_wrap(~ province, scale = "free_y") +
  theme_bw() +
  theme(legend.position = "bottom") +
  ggtitle("Number of vaccination children born in each province by month and year")
#ggsave(filename = file.path("..", "figures", "child_all_province_year.png"), width = 15, height = 10)
```

### Measle merge

```{r}
load(file.path(allp, "measle_all.Rdata"))

# Get public provinces only
measle_all <- measle_all[measle_all$province %in% publicProvinces &
                           measle_all$year >= 2017 & measle_all$year <= 2020,]
measle_all <- measle_all %>% distinct(., pid, province, district, commune, sex, dob, vacname2, vacdate, .keep_all = TRUE)
measle_all <- measle_all %>%
  group_by(pid, vacname2) %>%
  arrange(vacdate) %>%
  mutate(shot = 1:n()) %>%
  ungroup()
measle_all <- data.table(measle_all)
measle_all$age <- 2021 - measle_all$year
measle_all$sex <- ifelse(measle_all$sex==2, 1, measle_all$sex)
measle_all$sex <- factor(measle_all$sex, levels=c(0,1), labels=c("F","M"))
measle_all$shot2 <- ifelse(measle_all$shot >= 2, 2, measle_all$shot)
measle_all$start2 <- ifelse(measle_all$shot2 == 2, NA, measle_all$start)
measle_all$start2 <- ifelse(measle_all$vacname2 == "Measle_Mumps_Rubella" & measle_all$shot2 == 1, 12, measle_all$start2)
measle_all$vdelay2 <- measle_all$vagem2 - measle_all$start2
mea_all_overall <- measle_all[, .N, by = .(vyear, vmonth)]
mea_all <- measle_all[, .N, by = .(vyear, vmonth, type)]
```

## Measle vaccination

### Total measle vaccine shot

The table below describes the data of all Measle vaccination shots.
```{r}
t_all <- measle_all[year >= 2017 & year <= 2020] %>% select(age, sex, vacname2, shot2, vacdate, vdelay2, type) %>% 
  tbl_summary(
  # by=vdelay_outcome, # split table by group
  percent = "col",
  #missing = "no" # don't list missing data separately
  ) %>%
  add_n() %>% # add column with total number of non-missing observations
  # add_p(pvalue_fun=~style_pvalue(.x,digits=2)) %>% # test for a difference between groups
  modify_header(label = "**Variable**") %>% # update the column header
  bold_labels()
t_all
```

### Measle vaccination shots by month and year

Overall trend of Measle vaccination shots by month and year from 2017 to 2020 was shown in Figure 4. There was a steady increase in the total number of Measle vaccine shots from January to December, annually with a peak at different months in each year. There was a big reduction in total number of Measle vaccine shots in April, 2020 which was the first national lockdown period in Vietnam due to the COVID-19 pandemic (Figure 4). 

```{r}
ggplot(data = mea_all_overall[vyear >= 2017 & vyear <= 2020], aes(x = factor(vmonth), y = N/1000, color = factor(vyear))) +
  geom_rect(aes(xmin = 3, xmax = 5, ymin = -Inf, ymax = Inf), fill = "grey80", color = "grey90", alpha = 0.015) +
  geom_point(size = 3) +
  geom_line(aes(group = vyear), size = 1.5) +
  scale_x_discrete(name = "Month") +
  geom_smooth() +
  stat_smooth(method="loess", formula = y ~ x, colour="blue", size=1.5) +
  scale_y_continuous(name = "Number of shots (x 10^3)", breaks = seq(from = 0, to = 200, by = 50)) +
  scale_color_discrete(name = "Year") +
  theme_bw() +
  #facet_wrap(~ type) +
  ggtitle("Total number of Measle vaccination shots by month and year")
#ggsave(filename = file.path("..", "figures", "measle_all_overall_month_year.png"), width = 15, height = 10)
```
### Measle vaccination shots by type of clinics

Similar trends of Measle vaccine shots were observed for shots taken at public clinics only, private clinics only and for those taken vaccines at both public and private clinics, separately. In general, the number of Measle vaccine shots taken in public clinics were higher than those taken in private clinics (Figure 5).

```{r}
ggplot(data = mea_all[vyear >= 2017 & vyear <= 2020], aes(x = factor(vmonth), y = N/1000, color = factor(vyear))) +
  geom_rect(aes(xmin = 3, xmax = 5, ymin = -Inf, ymax = Inf), fill = "grey80", color = "grey90", alpha = 0.015) +
  geom_point(size = 3) +
  geom_line(aes(group = vyear), size = 1.5) +
  scale_x_discrete(name = "Month") +
  geom_smooth() +
  stat_smooth(method="loess", formula = y ~ x, colour="blue", size=1.5) +
  scale_y_continuous(name = "Number of shots (x 10^3)", breaks = seq(from = 0, to = 200, by = 50)) +
  scale_color_discrete(name = "Year") +
  theme_bw() +
  facet_wrap(~ type) +
  ggtitle("Total number of Measle vaccination shots by month and year")
#ggsave(filename = file.path("..", "figures", "measle_all_month_year.png"), width = 15, height = 10)
```
There has been a steady increase in total number of Measle vaccination shots from 2017 to the middle of 2020, then level off and decline in the last half of 2020 (Figure 6).

```{r}
ggplot(mea_all_overall[vyear >= 2017 & vyear <= 2020], aes(vmonth, N/1000, color = factor(vyear))) +
  geom_point(size=3) +
  geom_line(aes(group=vyear)) +
  geom_smooth () + 
  stat_smooth(method="loess", formula = y ~ x, colour="blue", size=1) +
  scale_x_discrete(name = "Month") +
  scale_y_continuous(name = "Number of shots (x 10^3)", breaks = seq(from = 0, to = 200, by = 50)) +
  scale_color_discrete(name = "Year") +
  theme_bw() +
  ggtitle("Total number of Measle vaccination shots by month and year") +
  facet_wrap(~ vyear, scale = "free_y")
#ggsave(filename = file.path("..", "figures", "measle_all_overall_month_year.png"), width = 15, height = 10)
```

### Delays in vaccination

We quantifed the delays in Measle vaccination over years. **We use the first dose for this analysis. Delay is defined as the actual date of vaccination later than at least 1 month compared to the expected date**. The distribution of the delays in Measle vaccination (in months) for children born from 2017 to 2020 was shown in Figure 7. The earlier the children born, the longer in delays of Measle vaccination they could probably have. The longest delays of Measle vaccination can be up to more than 30 months for the Measle vaccine and 30 months for the Measle-Rubella vaccine compared to the schedule of those vaccines they supposed to take. According the recommended immunization schedule, the schedule of Measle and Measle-Rubella vaccines taken was 9 months and 18 months after the children were born, respectively. There was no data on the delays of Measle vaccination from the last haft of 2019 to the end of 2020 (Figure 7).

```{r}
measle_all <- measle_all[measle_all$shot == 1,]
measle_all$expect_vdate <- measle_all$dob %m+% months(measle_all$start)
measle_all$expect_vmonth <- month(measle_all$expect_vdate)
measle_all$expect_vyear <- year(measle_all$expect_vdate)

ggplot(data = measle_all[year >= 2017 & year <= 2020 & vdelay>=0,], aes(x = vdelay, y = factor(expect_vmonth), color = factor(expect_vyear))) +
  geom_density_ridges() +
  theme_ridges() + 
  theme(legend.position = "none") +
  facet_grid(vacname2~expect_vyear, scale = "free_y") +
  labs(y = "Expected month of vaccination", x="Duration of delays in Measle vaccination (months)") +
  scale_color_discrete(name = "Year") +
  ggtitle("Delays in Measle vaccination (months) by expected month and year")
#ggsave(filename = file.path("..", "figures", "measle_all_delay_year.png"), width = 15, height = 10)
```

```{r}
#load(file.path("..", "..", "tuann349_vad", "vacinfo.Rdata"))
#load(file.path("..", "..", "tuann349_vad", "vacinfo_private.Rdata"))
library(lubridate)

measle_all_delay <- measle_all %>%
  select(pid, province, district, commune, sex, dob, vacname2, start2, shot2, vacdate, vyear, vmonth, year, month, vagem, vagem2, vdelay2, vsche, vsyear, vsmonth, vdelayd, type, type2) %>% filter(year>=2017 & year<=2020) %>%
  mutate(vdelay_outcome=ifelse(vdelay2>=1,1,0), # vdelay >= 1 month
         age=2021-year,
         shot2=factor(shot2),
         vacdate_f=ifelse(vacdate<"2020-04-01",0,1),
         vacdate_f=factor(vacdate_f, levels = c(0,1), labels = c("Before", "After"))) %>% 
  filter(province!="Tinh tap huan")

measle_all_delay$vdelay_outcome <- factor(measle_all_delay$vdelay_outcome, levels = c(0, 1),
                                          labels = c("Not delay", "Delay"))
# Check how many provinces in measle_all_delay
# length(unique(measle_all_delay$province))
```

### Factors associated with delays in vaccination

The probability of the delays in Measle vaccination (dichotomous outcome: delay versus no delay) was modeled to see if it depends on other related factors. The delays in Measle vaccination were shown to be dependent on age (p<0.001), province (p<0.001), type of vaccine (p<0.001), vaccination period (before and after COVID-19 pandemic) (p<0.001).

```{r}
t3 <- measle_all_delay %>% select(age, sex, vacname2, vacdate_f, vdelay2, vdelay_outcome, type) %>% 
  tbl_summary(
  by=vdelay_outcome, # split table by group
  percent = "row",
  #missing = "no" # don't list missing data separately
  ) %>%
  add_n() %>% # add column with total number of non-missing observations
  add_p(pvalue_fun=~style_pvalue(.x,digits=2)) %>% # test for a difference between groups
  modify_header(label = "**Variable**") %>% # update the column header
  bold_labels()
```

```{r}
t3
```

```{r}
#install.packages("rms")
library("rms")
theme_set(theme_bw())
dd <- datadist(measle_all_delay)
options(datadist="dd")
dd

fit1 <- lrm(vdelay_outcome~age, data=measle_all_delay, maxit=1000)
anova(fit1)
pred <- Predict(fit1, age=1:4, fun=plogis)
ggplot(pred, ylab=expression(paste(hat(P),"(delay)")), ylim=c(0,1)) +
  #scale_x_discrete(name = "Age") + 
  theme(text = element_text(size=15), axis.text.x = element_text(size=14), axis.text.y = element_text(size=14), legend.text=element_text(size=14)) +
  geom_line(size=1.25) + 
  geom_ribbon(aes(ymin = pred$lower, ymax = pred$upper), alpha=0.05) + 
  xlab("Age") +
  annotate("text", x=3.5, y=0.8, label=paste0("p-value<0.001"), parse = TRUE)
#ggsave(filename = file.path("..", "figures", "measle_all_delay_age.png"), width = 15, height = 10)
```

```{r}
fit2 <- lrm(vdelay_outcome~sex, data=measle_all_delay, maxit=1000)
anova(fit2)
pred <- Predict(fit2, sex=NA, fun=plogis)

ggplot(pred, ylab=expression(paste(hat(P),"(delay)"))) +
  theme(text = element_text(size=15), axis.text.x = element_text(size=14), axis.text.y = element_text(size=14), legend.text=element_text(size=14)) +
  geom_line(size=1.25) + 
  geom_ribbon(aes(ymin = pred$lower, ymax = pred$upper), alpha=0.05) + 
  xlab("Sex") +
  annotate("text", x=0.9447, y=0.5, label=round(anova(fit2)[5],3), parse = TRUE)
```

```{r}
fit3 <- lrm(vdelay_outcome~province, data=measle_all_delay, maxit=1000)
anova(fit3)
pred <- Predict(fit3, province=NA, fun=plogis)
ggplot(pred, ylab=expression(paste(hat(P),"(delay)"))) +
  theme(text = element_text(size=10), axis.text.x = element_text(size=10), axis.text.y = element_text(size=6), legend.text=element_text(size=10)) +
  geom_line(size=1.25) + 
  geom_ribbon(aes(ymin = pred$lower, ymax = pred$upper), alpha=0.05) + 
  xlab("Province") +
  annotate("text", x=25, y=0.2, label=paste0("p-value<0.001"), parse = TRUE)
#ggsave(filename = file.path("..", "figures", "measle_all_delay_province.png"), width = 15, height = 10)
```

```{r}
fit4 <- lrm(vdelay_outcome~vacdate_f, data=measle_all_delay, maxit=1000)
anova(fit4)
pred <- Predict(fit4, vacdate_f=NA, fun=plogis)
ggplot(pred, ylab=expression(paste(hat(P),"(delay)"))) +
  theme(text = element_text(size=10), axis.text.x = element_text(size=10), axis.text.y = element_text(size=10), legend.text=element_text(size=10)) +
  geom_line(size=1.25) + 
  geom_ribbon(aes(ymin = pred$lower, ymax = pred$upper), alpha=0.05) + 
  xlab("Vaccination Period") +
  annotate("text", x=2, y=0.95, label=paste0("p-value<0.001"), parse = TRUE)
#ggsave(filename = file.path("..", "figures", "measle_all_delay_period.png"), width = 15, height = 10)
```

```{r}
fit5 <- lrm(vdelay_outcome~type, data=measle_all_delay, maxit=1000)
anova(fit5)
pred <- Predict(fit5, fun=plogis)
ggplot(pred, ylab=expression(paste(hat(P),"(delay)"))) +
  theme(text = element_text(size=10), axis.text.x = element_text(size=10), axis.text.y = element_text(size=10), legend.text=element_text(size=10)) +
  geom_line(size=1.25) + 
  geom_ribbon(aes(ymin = pred$lower, ymax = pred$upper), alpha=0.05) + 
  xlab("Type of vaccination") +
  annotate("text", x=2, y=0.935, label=paste0("p-value<0.001"), parse = TRUE)
#ggsave(filename = file.path("..", "figures", "measle_all_delay_type.png"), width = 15, height = 10)
```

```{r}
# fit6 <- lrm(vdelay_outcome~vacdate_f + province + age + type, data=measle_all_delay, maxit=1000)
# anova(fit6)
# pred <- Predict(fit6, fun=plogis)
# ggplot(pred[1:2,5:7], ylab=expression(paste(hat(P),"(delay)"))) +
#   theme(text = element_text(size=10), axis.text.x = element_text(size=10), axis.text.y = element_text(size=10), legend.text=element_text(size=10)) +
#   geom_line(size=1.25) + 
#   geom_ribbon(aes(ymin = min(pred$lower), ymax = max(pred$upper)), alpha=0.05) + 
#   xlab("Type of vaccination") +
#   annotate("text", x=2, y=0.935, label=paste0("p-value<0.001"), parse = TRUE)
```
